feat: json file streaming implementation
